Initial commit
This is the initial commit of the Ardash dApp, which can be used to
assist with the pre-processing and uploading of MPEG-DASH videos to the decentralized
Arweave permaweb.

This tool is composed of 3 parts:

* cli: Command-line tools that allow you to convert a H.264/MP4 video
into MPEG-DASH segments, and upload them into the permaweb. It relies on
gpac's MP4Box for the encoding of videos;

* backend: Tiny flask server that allows you to wrap command-line
utilies, and pre-process videos as a service;

* dApp: client-side distributed application. Connects to the backend
for video pre-processing, and allows you to deploy videos by relying
on arweave-js.